refactor(bulk-editor): extract the appearance content component
Moves the tab bar, tab panels, and the active field set wiring out of
the app into a BulkEditorContent component, per review feedback. The
state stays in the store: the premium bundle's fills (AI buttons,
smart-select) read and dispatch the same view state.
